backend/ccxt-server/pyproject.toml
[project]
name = "ccxt-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ccxt>=4.5.32",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "hvac>=2.4.0",
    "protobuf>=6.33.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]
